Prevent implementation errors with multi-tenant key providers
The intent of the injectTenantMetadata() method is to return either the original encrypted row -OR- with additional fields. If a user returns fewer fields, that's very likely an error condition that needs to be caught.

This check is added to ensure, by default, these failures are caught and no data is lost. It can, however, be turned off.